requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
Babel==2.4.0
braintree==3.37.2
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
cryptography==1.9
Django==1.11.1
django-bootstrap-form==3.2.1
django-cart==1.0.4
django-countries==4.5
django-phonenumber-field==1.3.0
gunicorn==19.7.1
idna==2.5
olefile==0.44
packaging==16.8
phonenumberslite==8.5.2
Pillow==4.1.1
pycparser==2.17
PyJWT==1.5.0
pyparsing==2.2.0
pytz==2017.2
requests==2.17.3
six==1.10.0
stripe==1.59.0
suds-jurko==0.6
urllib3==1.21.1
whitenoise==3.3.0
xmltodict==0.11.0
